optimized cvtsi2ss in asm

+++ JAProjekt/JAProjektCppLibrary/JAProjektCppLibrary.cpp
@@ -18,7 +18,7 @@ void cppBinarization1(char* begin, char* end, long biWidth, float treshold)
 		float R = (unsigned char)*(currPos + 2);
 		float G = (unsigned char)*(currPos + 1);
 		float B = (unsigned char)*(currPos + 0);
-		uint8_t Res = R * 0.299 + G * 0.587 + B * 0.144;
+		uint16_t Res = R * 0.299 + G * 0.587 + B * 0.144;
 		if (Res > 256 * treshold)
 		{
 			*(currPos) = 255;
